Flyttet endepunkt til ny controller
