+++ src/main/kotlin/no/nav/tilleggsstonader/sak/migrering/routing/SøknadRoutingController.kt
@@ -2,8 +2,6 @@ package no.nav.tilleggsstonader.sak.migrering.routing
 
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import no.nav.tilleggsstonader.kontrakter.felles.IdentStønadstype
-import no.nav.tilleggsstonader.sak.behandling.domain.BehandlingStatus
-import no.nav.tilleggsstonader.sak.fagsak.FagsakService
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvisIkke
 import no.nav.tilleggsstonader.sak.infrastruktur.sikkerhet.EksternApplikasjon
 import no.nav.tilleggsstonader.sak.infrastruktur.sikkerhet.SikkerhetContext
@@ -14,39 +12,18 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
-@RequestMapping(path = ["/api/ekstern"])
+@RequestMapping(path = ["/api/ekstern/routing-soknad"])
 @ProtectedWithClaims(issuer = "azuread")
 class SøknadRoutingController(
     private val søknadRoutingService: SøknadRoutingService,
-    private val fagsakService: FagsakService,
 ) {
 
-    @PostMapping("routing-soknad")
+    @PostMapping()
     @ProtectedWithClaims(issuer = "azuread", claimMap = ["roles=access_as_application"])
     fun sjekkRoutingForPerson(@RequestBody request: IdentStønadstype): SøknadRoutingResponse {
         feilHvisIkke(SikkerhetContext.kallKommerFra(EksternApplikasjon.SOKNAD_API), HttpStatus.UNAUTHORIZED) {
             "Kallet utføres ikke av en autorisert klient"
         }
         return søknadRoutingService.sjekkRoutingForPerson(request)
-    }
-
-    @PostMapping("harBehandling")
-    @ProtectedWithClaims(issuer = "azuread", claimMap = ["roles=access_as_application"])
-    fun hentBehandlingStatusForPersonMedStønadstype(@RequestBody identStønadstype: IdentStønadstype): Boolean {
-        val behandlinger = fagsakService.hentBehandlingerForPersonOgStønadstype(
-            identStønadstype.ident,
-            identStønadstype.stønadstype,
-        )
-        if (behandlinger.isNotEmpty()) {
-            val validStatuses = listOf(
-                BehandlingStatus.OPPRETTET,
-                BehandlingStatus.UTREDES,
-                BehandlingStatus.FATTER_VEDTAK,
-                BehandlingStatus.SATT_PÅ_VENT,
-            )
-            val behandlingStatus = behandlinger.map { it.status }
-            return behandlingStatus.any { it in validStatuses }
-        }
-        return false
     }
 }
